requirements.txt
telethon
rich
pyfiglet
Pillow